Changed how different needed files for the cases are copied
Changing the name of the project would make a reference error due to
relative paths. To solve this the files needed to movements, accinputs,
etc. are now copied both to the case root and the out folder. If the out
folder gets removed of its name changes, the files on the project root
will be copied instead of having to search for the file full path again.


Former-commit-id: 371681bcd3fa746f5e1e9aa5e1c97572b43b46e6