Refactor: Structure and cleaning code
Now all the individual classes are separed
in different files so linting is faster and more
reliable.

Removed wildcard imports and made it easier
for editors to spot problems and sketchy things.

Also removed pieces of bad python practices
around the code, like list comprehensions for
iterative tasks or sequence size cheking for
emptyness.

That all was among other things.