MKBasedProperties: Add real mk abstraction
This was a bit of a mess. For some reason I forgot
that mkbound 0 is not the same as mkfluid 0, but
the application just gave the user whatever mk
properties were configured for "0" whether it was
fluid or bound.

In DualSPHysics the "real" mk is basically the same
number for mkfluids (mkfluid 1 is realmk 1) but for
bounds it begins at 11 (mkbound 3 is realmk 14).

What this means basically is that realmks [0, 10] are
fluids while [11, ...) are bounds.

This is fixed now and also the mk properties are now
created on demand.


Former-commit-id: 2a9df4574a5f6a0a4723363c739d5840760c47e2