Movements: Fixes redundant recursion etc
In the previous implementation of movements each
motion had a reference to its parent. When saving
and trying to recursively converts objects and its
childs to dictionaries, the redundant upwards
reference was causing infinite recursion, so we got
rid of that.

Basically that was used only to figure out which
movement list to update when changing or deleting
a motion, so that is now achieved saving a reference
to the currently selected movement in the
MovementDialog.


Former-commit-id: 6d67e5e7d167f0f7091c2a9add085f52cd2e6de0