SimulationWidget: totalpartsout is now initialized
There was an error at simulation where if no parts
were out at the initial time step computed by
DualSPHysics it would throw an exeception because
the variable didn't exist (yet).

So basically this is a fix for that initializing the variables
to the base values so it won't happen.


Former-commit-id: a513dc9b3050d16187911624caac92c4a01c0b79