WavepaddlesRenderer: Change key existence assumption and fix AWAS
In the render method of the WavePaddlesRenderer there
was an assumption of the key "generator" existing in a
movement, which not always happens, specifically if there's
only one movement in the case and it is a regular movement.

Also, fixes AWAS configuration being written onto the XML even
if it was not enabled.


Former-commit-id: 4eeaaf46eacb15eafd2eeb31365624c53808a859