Moorings: Prevent loss of settings for bodies that were already selected
When pressing the button to configure the MoorDyn parameters the
list of bodies to configure there was constructed from scratch fom the
previously selected floating body list. That caused all the bodies to
loose configuration when entering that window.

Now the configuration is lost only if the user unchecks some floating
body and then presses the button.


Former-commit-id: 14b9b80ff37e0bebccaa48ea3d78e0f93da08503