MoorDyn: Line configuration now reads appropriate bodies
Before this change MoorDyn would only display bodies to reference
a vessel connection stored in the main case structure. However when
you're configuring Moorings a temporal strucutre is created so only
when you press OK it gets saved.

Basically that meant that if you deselected some body previously
selected the Line configuration would still show it, or if it were the
first time you configured moorings for that case the Line configuration
would not show anything


Former-commit-id: debb2a5b844f29da0f319e20ec32005af9c2c3b1