Simulation: Save the case at simulation finish
This changes accounts for most of users don't realizing they have to
save the case to preserve the current state of knowledge about the case.

When you finish a simulation, the UI reacts to it accordingly enabling
and disabling widgets, but if you don't save the case that decision is
not stored on disk, meaning that if you reopen the case the previous
state of the widgets were not saved.

Just a mild inconvenience but also a quick fix.


Former-commit-id: 6fd30c425ed726fb8a781a455e5dea946075f748