InformationDialog: Adds HTML/Plain text mode attribute
This patch changes the way how the details text is feeded into the
QTextEdit object, as passing it by the constructor lead to Qt trying
to figure out if it is plain text or HTML, which is wrong on the output
of most of the DualSPHysics tools.

This way you can just choose if you prefer plain text or HTML, using
the InformationDetailsMode enum.


Former-commit-id: 1eb59ab60713e8f1b3305e787ce813d2649379bb