ApplicationSettings: Changes pickle for JSON serialization
Pickling this object just seemed a bit overkill for what this class
is used for. Now it serializes to JSON and loads it checking which
properties does exist and which do not, so in the future we don't
cause any incompatibilities.

It also saves a "json_version" key which will be useful in the future
if for some reason we want to break compatibility and need to check
in some way if the JSON is from an old version or not