InletConfigDialog: Prevent buttons being too short in height
In certain resolutions and screen configurations the buttons on
each of the InletZoneWidget(s) were too short, causin the text to
overflow, resulting on texts that were difficult to read.

This patch simply adds a bit of height on each of the cells in the
table inside the InletConfigDialog to prevent that.