Autofill+related: Changes the way we check for autofillable objects
Before this patch each one of the externally added objects had the
autofill property set either to True or False, not None. So we used
that to check if an object was imported before. However this is not
a tight check as None is not a value supposed to be on a Boolean.

This patch changes this logic to set the internal names of the
imported objects to begin with "external_" and having the Label be
the one the user had input in the first place. This means that now
in the XML the object name will be written as external_{name} but I
suppose we can live with it. Maybe in the future we should change
whats written to the XML to be the Label, not the Name.